+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 6.3
 - SQL add Donate link in rosariosis.sql, rosariosis_es.sql & rosariosis_fr.sql
 - HTML smaller size for 'Displaying %d through %d' in ListOutput.fnc.php
 - Create Student Account: Fix reload page on School change, no AJAX in General_Info.inc.php
+- SQL add CREATE_STUDENT_ACCOUNT_DEFAULT_SCHOOL to CONFIG table in rosariosis.sql
 
 Changes in 6.2.3
 ----------------